CSS HTML index and JS folder
Added a main.js file to js folder - it's empty so far.

Neatened the CSS file by listing elements to style in the order they appear on the page.

Adjusted lots of margins and padding of the sections.

There's a bug with the scroll-spy. It starts of fine going down the page however when you return to the top it assumes that jumbotron is the artists section. This may be because it uses the bottom of sections as a position cue.

The index.html has:

-text-area element for the message input box to allow more text and to "wrap" the input text.

Given the artist section rows better IDs.